Refactor app into modular domain, service, UI, and init layers
