+++ index.html
@@ -203,6 +203,11 @@
 
     </div>
 
+    <script src="src/domain/insights.js"></script>
+    <script src="src/domain/tasks.js"></script>
+    <script src="src/services/storage.js"></script>
+    <script src="src/ui/render.js"></script>
+    <script src="src/app/init.js"></script>
     <script src="script.js"></script>
 
 </body>
